Alura/Django_rest_4.0/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
defusedxml==0.7.1
Django==4.1.7
django-admin-honeypot==1.1.0
django-cors-headers==3.14.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
Faker==17.5.0
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
validate-docbr==1.10.0
